services/pathway-processor/requirements.txt
python-dotenv
numpy
openai
tiktoken
pandas